# openapi.yaml
openapi: 3.0.3
info:
  title: Jelou B2B Backoffice API
  description: API consolidada para gestión de clientes, productos, órdenes y orquestación
  version: 1.0.0
  contact:
    name: API Support

paths:
  # ==================== Orchestrator API ====================
  /orchestrator/create-and-confirm-order:
    servers:
      - url: https://63efqt72we.execute-api.us-east-1.amazonaws.com
        description: Lambda Orchestrator (Production)
    post:
      summary: Create and confirm order (orchestrator)
      description: Orchestrates the complete flow of validating customer, creating order, and confirming it in a single operation
      tags:
        - Orchestrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrchestratorRequest'
            examples:
              with_correlation:
                summary: With correlation ID
                value:
                  customer_id: 1
                  items:
                    - product_id: 2
                      qty: 3
                  idempotency_key: unique-key-123
                  correlation_id: req-abc-123
              without_correlation:
                summary: Without correlation ID
                value:
                  customer_id: 1
                  items:
                    - product_id: 2
                      qty: 3
                  idempotency_key: unique-key-456
      responses:
        '201':
          description: Order created and confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrchestratorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrchestratorErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrchestratorErrorResponse'

  # ==================== Customers API ====================
  /customers:
    servers:
      - url: http://localhost:3001
        description: Customers API
    post:
      summary: Create customer
      tags:
        - Customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Search customers
      tags:
        - Customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search term for name, email, or phone
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination (last ID)
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Number of results per page
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  /customers/{id}:
    servers:
      - url: http://localhost:3001
        description: Customers API
    get:
      summary: Get customer by ID
      tags:
        - Customers
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update customer
      tags:
        - Customers
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete customer
      tags:
        - Customers
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Customer deleted successfully
        '404':
          description: Customer not found

  /internal/customers/{id}:
    servers:
      - url: http://localhost:3001
        description: Customers API
    get:
      summary: Get customer by ID (internal service endpoint)
      tags:
        - Customers
      security:
        - ServiceToken: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          description: Missing or invalid authorization
        '404':
          description: Customer not found

  # ==================== Products API ====================
  /products:
    servers:
      - url: http://localhost:3002
        description: Orders API
    post:
      summary: Create product
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error or SKU already exists

    get:
      summary: Search products
      tags:
        - Products
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by SKU or name
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  /products/{id}:
    servers:
      - url: http://localhost:3002
        description: Orders API
    get:
      summary: Get product by ID
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

    patch:
      summary: Update product (price/stock)
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found

  # ==================== Orders API ====================
  /orders:
    servers:
      - url: http://localhost:3002
        description: Orders API
    post:
      summary: Create order
      description: Validates customer, checks stock, creates order (CREATED), and decrements stock in a transaction
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              simple:
                value:
                  customer_id: 1
                  items:
                    - product_id: 2
                      qty: 3
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithItems'
        '400':
          description: Invalid customer or insufficient stock

    get:
      summary: Search orders
      tags:
        - Orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [CREATED, CONFIRMED, CANCELED]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Filter orders from this date
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: Filter orders until this date
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  /orders/{id}:
    servers:
      - url: http://localhost:3002
        description: Orders API
    get:
      summary: Get order with items
      tags:
        - Orders
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithItems'
        '404':
          description: Order not found

  /orders/{id}/confirm:
    servers:
      - url: http://localhost:3002
        description: Orders API
    post:
      summary: Confirm order (IDEMPOTENT)
      description: Confirms an order using X-Idempotency-Key header. Retries with same key return cached response.
      tags:
        - Orders
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Order confirmed (or already confirmed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithItems'
        '400':
          description: Missing idempotency key or invalid state
        '404':
          description: Order not found

  /orders/{id}/cancel:
    servers:
      - url: http://localhost:3002
        description: Orders API
    post:
      summary: Cancel order and restore stock
      description: |
        Cancels order and restores stock:
        - CREATED: Always cancels
        - CONFIRMED: Cancels only if less than 10 minutes have passed
      tags:
        - Orders
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWithItems'
        '400':
          description: Cannot cancel (confirmed order older than 10 minutes)
        '404':
          description: Order not found

components:
  parameters:
    CustomerId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Customer ID

    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Product ID

    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Order ID

    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
      description: Unique key for idempotent operations
      example: unique-key-abc-123

  schemas:
    # ==================== Orchestrator Schemas ====================
    OrchestratorRequest:
      type: object
      required:
        - customer_id
        - items
        - idempotency_key
      properties:
        customer_id:
          type: integer
          minimum: 1
          example: 1
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - product_id
              - qty
            properties:
              product_id:
                type: integer
                minimum: 1
                example: 2
              qty:
                type: integer
                minimum: 1
                example: 3
        idempotency_key:
          type: string
          minLength: 1
          example: unique-key-abc-123
        correlation_id:
          type: string
          example: req-abc-123

    OrchestratorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        correlationId:
          type: string
          example: req-abc-123
        data:
          type: object
          properties:
            customer:
              $ref: '#/components/schemas/Customer'
            order:
              $ref: '#/components/schemas/OrderWithItems'

    OrchestratorErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        correlationId:
          type: string
          example: req-abc-123
        error:
          type: string
          example: Validation error
        message:
          type: string
          example: "customer_id: Expected number, received string"

    # ==================== Customer Schemas ====================
    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: ACME Corporation
        email:
          type: string
          format: email
          example: ops@acme.com
        phone:
          type: string
          nullable: true
          example: +1-555-0100
        created_at:
          type: string
          format: date-time
          example: '2025-11-21T12:00:00.000Z'

    CreateCustomerRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: ACME Corporation
        email:
          type: string
          format: email
          maxLength: 255
          example: ops@acme.com
        phone:
          type: string
          maxLength: 50
          example: +1-555-0100

    UpdateCustomerRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        phone:
          type: string
          maxLength: 50

    # ==================== Product Schemas ====================
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        sku:
          type: string
          example: LAPTOP-PRO-15
        name:
          type: string
          example: Laptop Pro 15 inch
        price_cents:
          type: integer
          description: Price in cents
          example: 129900
        stock:
          type: integer
          example: 50
        created_at:
          type: string
          format: date-time

    CreateProductRequest:
      type: object
      required:
        - sku
        - name
        - price_cents
      properties:
        sku:
          type: string
          minLength: 1
          maxLength: 255
          example: LAPTOP-PRO-15
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Laptop Pro 15 inch
        price_cents:
          type: integer
          minimum: 0
          example: 129900
        stock:
          type: integer
          minimum: 0
          default: 0
          example: 50

    UpdateProductRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        price_cents:
          type: integer
          minimum: 0
        stock:
          type: integer
          minimum: 0

    # ==================== Order Schemas ====================
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        customer_id:
          type: integer
          example: 1
        status:
          type: string
          enum: [CREATED, CONFIRMED, CANCELED]
          example: CREATED
        total_cents:
          type: integer
          example: 389700
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
          nullable: true
        canceled_at:
          type: string
          format: date-time
          nullable: true

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        product_id:
          type: integer
        qty:
          type: integer
        unit_price_cents:
          type: integer
        subtotal_cents:
          type: integer

    OrderWithItems:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'

    CreateOrderRequest:
      type: object
      required:
        - customer_id
        - items
      properties:
        customer_id:
          type: integer
          minimum: 1
          example: 1
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - product_id
              - qty
            properties:
              product_id:
                type: integer
                minimum: 1
                example: 2
              qty:
                type: integer
                minimum: 1
                example: 3

    # ==================== Shared Schemas ====================
    Pagination:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
          example: '10'
        hasMore:
          type: boolean
          example: true
        limit:
          type: integer
          example: 20

    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        message:
          type: string
          example: Invalid input
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string

  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: Service-to-service authentication token

tags:
  - name: Orchestrator
    description: Lambda orchestrator endpoints for complex workflows
  - name: Customers
    description: Customer management endpoints
  - name: Products
    description: Product management endpoints
  - name: Orders
    description: Order management endpoints